+++ src/main/java/it/pagopa/ecommerce/payment/methods/infrastructure/rule/AmountTypeLangFilterRule.java
@@ -5,8 +5,6 @@ import it.pagopa.ecommerce.payment.methods.infrastructure.PspRepository;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Flux;
 
-import java.math.BigInteger;
-
 @Component
 public class AmountTypeLangFilterRule implements IFilterRule {
     @Override
@@ -25,6 +23,6 @@ public class AmountTypeLangFilterRule implements IFilterRule {
                                      String language,
                                      String paymentTypeCode
     ) {
-        return pspRepository.findPspMatchAmountTypeLang(BigInteger.valueOf(amount), paymentTypeCode, language);
+        return pspRepository.findPspMatchAmountTypeLang(amount, paymentTypeCode, language);
     }
 }
